NCKU資安新秀/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
csp==0.0.0
Django==4.2.7
django-cors-headers==4.3.0
django-csp==3.7
django-scheduler==0.10.1
djangorestframework==3.14.0
icalendar==5.0.11
idna==3.4
numpy==1.26.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
